Make the QMK sculpt compatible with the simplified LED driving circuit of proposed rev5 board design.
Rev5 board design removes the charge pump.

When the common pin for both LEDs is driven with a 50% duty-cycle rectangular signal both LEDs can be used without additional circuit.
QMK has a feature to change the brightness of background LEDs and we are just going to use this feature to produce such a signal.